Move dashboard containers
