+++ src/components/containers/Dashboard/Receipts/Receipts.css
