+++ src/components/containers/Dashboard/Receipts/Receipts.spec.js
