+++ src/components/containers/Dashboard/UserDetails/UserDetails.spec.js
